backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
httpx==0.27.2
sqlalchemy==2.0.36
pydantic==2.10.0
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.3
resend==2.6.0
python-dotenv==1.0.1